rename validators to checkers
