+++ .script/utils/jsonSchemaChecker.ts
